# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Happening API (Evently)
  version: 1.0.0
  description: |
    OpenAPI specification for the Happening/Evently backend. 
    This API demonstrates a scalable, production-grade architecture featuring:
    - **Redis Caching:** For all public GET routes to handle read-heavy traffic.
    - **RabbitMQ Queueing:** For the /bookings endpoint to handle write-heavy "flash sale" scenarios.
    - **PostgreSQL Transactions (FOR UPDATE):** To ensure data integrity and prevent overselling.
    - **Full Admin & Host Management:** Complete with approval flows and analytics.
servers:
  - url: http://localhost:5000/api
    description: Local development server
  - url: https://your-deployed-url.com/api
    description: Production server

# ==========================================================
#  Reusable Components (Schemas & Security)
# ==========================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "User's JWT token, prefixed with 'Bearer '"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      example:
        error: "Invalid or expired token"
...
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        venue:
          type: string
        date_time:
          type: string
          format: date-time
        total_tickets:
          type: integer
        available_tickets:
          type: integer
        price:
          type: number
          format: float
        category:
          type: string
          enum: [music, food, sports, tech, arts, health, workshop]
        image_url:
          type: string
          format: uri
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        organizer_name:
          type: string
      example:
        id: "e1f2a3b4-..."
        title: "Live Tech Conference"
        description: "The biggest tech conference of the year."
        venue: "1 Infinite Loop, Cupertino, CA 95014"
        date_time: "2025-12-01T20:00:00Z"
        total_tickets: 1000
        available_tickets: 450
        price: 999.0
        category: "tech"
        image_url: "https://example.com/image.jpg"
        latitude: 37.3318
        longitude: -122.0312
        organizer_name: "Admin"
...
paths:
...
  /events/public/search:
    get:
      summary: Search public events (cached)
      tags: [Events]
      parameters:
        - { name: query, in: query, schema: { type: string }, description: "Search query text" }
        - { name: category, in: query, schema: { type: string }, description: "Event category" }
        - { name: min_price, in: query, schema: { type: number }, description: "Minimum price" }
        - { name: max_price, in: query, schema: { type: number }, description: "Maximum price" }
        - { name: location, in: query, schema: { type: string }, description: "Location name (e.g., 'Bhopal')" }
        - { name: radius, in: query, schema: { type: integer }, description: "Radius in miles" }
        - { name: date, in: query, schema: { type: string, enum: [today, tomorrow, weekend, week] }, description: "Date range filter" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      $ref: "#/components/schemas/Event"
...
  /events/{id}:
    get:
      summary: Get a single event by ID (cached)
      tags: [Events]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  event:
                    $ref: "#/components/schemas/Event"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    
    put:
      summary: Update an event (Host or Admin)
      tags: [Events]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Event"
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "403":
          description: Forbidden (Not event owner)
        "404":
          description: Not Found
  # --- END OF FIX ---

  # --- Protected Event Routes (Host/Admin) ---
  /events:
    post:
      summary: Create a new event (Admin or Approved Host)
      tags: [Events]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Event" # Simplified, re-uses Event schema
      responses:
        "201":
          description: Created (or pending approval)
        "401":
          description: Unauthorized
        "403":
          description: Forbidden (Not an approved host)
  /events/{id}/cancel:
    post:
      summary: Cancel an event & refund all bookings (Host or Admin)
      tags: [Events]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "403":
          description: Forbidden (Not event owner)
        "404":
          description: Not Found

  # --- Protected User Profile ---
...
  /events/admin/pending:
    get:
      summary: (Admin) Get all pending events for review
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
...